Fold "start from your brand" into the Company name field
Replace the standalone BrandImport section with BrandNameField: the
Company name input now carries a quiet "From website" affordance that
morphs the same field into a URL importer, pulling the company name and
brand color, then morphs back with the site favicon as a leading marker,
a color dot that re-opens the detected palette, and a brand-color pulse.

Built on one persistent input whose bindings switch by mode, with
enter-only keyed transitions instead of AnimatePresence exits, so the
morph keeps focus and caret and never wedges.
